blocks now stored internally as uint16 to reduce ram count

+++ include/Chunk.hpp
@@ -23,6 +23,8 @@ class Chunk {
         // absolute world height -- callers evaluating a density function need
         // world coordinates for continuity across chunk boundaries anyway, so
         // this avoids a separate local/world Y conversion at every call site.
+        // Public interface stays Int32 (matches the wire format's block-state
+        // field width); internal storage is narrower -- see _blocks.
         void setBlock(int localX, int worldY, int localZ, Int32 blockStateId);
         Int32 getBlock(int localX, int worldY, int localZ) const;
 
@@ -38,7 +40,11 @@ class Chunk {
         int index(int localX, int worldY, int localZ) const;
         int _chunkX;
         int _chunkZ;
-        std::array<Int32, SECTION_COUNT * 4096> _blocks{}; // defaults to 0 (air)
+        // Int16, not Int32: real vanilla 1.21 has ~27,000 block states total,
+        // comfortably under the 32,767 signed-16-bit ceiling, so this isn't a
+        // shortcut specific to today's 4-block palette -- it's safe for the
+        // entire real game. Cuts this array (2/3 of a Chunk's size) in half.
+        std::array<Int16, SECTION_COUNT * 4096> _blocks{}; // defaults to 0 (air)
         std::array<uint8_t, SECTION_COUNT * 4096> _skyLight{};
         std::array<uint8_t, SECTION_COUNT * 4096> _blockLight{};
         Int32 _biomeId = 0;
